requirements.txt
Django==2.2.1
celery==4.3.0
django-celery-results==1.1.2
xlrd==1.2.0
xlwt==1.3.0
lxml==4.4.1
djangorestframework==3.9.4
requests==2.12.1